% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_spline_data.R
\name{analyze_boot_splines}
\alias{analyze_boot_splines}
\alias{analyze_boot_splines.boot_splines_data}
\title{Estimate confidence intervals for bootstrapped splines data}
\usage{
analyze_boot_splines(data)

\method{analyze_boot_splines}{boot_splines_data}(data)
}
\arguments{
\item{data}{The output of the \code{boot_splines_data} function}
}
\value{
A dataframe indicating means and CIs for each time-bin
}
\description{
Estimates a confidence interval over the difference between means (within- or between-subjects)
from a \code{boot_splines_data} object. Confidence intervals are derived from the alpha
used to shape the dataset (e.g., alpha = .05, CI=(.025,.975); alpha=.01, CI=(.005,.0995))
}
\section{Methods (by class)}{
\itemize{
\item \code{boot_splines_data}: 
}}

