% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_sequence_data.R
\name{analyze_time_bins}
\alias{analyze_time_bins}
\title{analyze_time_bins()}
\usage{
analyze_time_bins(data, ...)
}
\arguments{
\item{data}{The output of the 'make_time_sequence_data' function}

\item{...}{Any other arguments to be passed to the selected 'test' function (e.g., paired,
var.equal, etc.)}

\item{data_options}{}

\item{predictor_column}{The variable whose test statistic you are interested in}

\item{test}{What type of test should be performed in each time bin? Supports t.test, wilcox,
lm, or lmer.}

\item{threshold}{Value of statistic used in determining significance}

\item{alpha}{Alpha value for determining significance, ignored if threshold is given}

\item{formula}{What formula should be used for the test? Optional for all but lmer, if unset will
use \code{Prop ~ [predictor_column]}}

\item{return_model}{In the returned dataframe, should a model be given for each time bin, or just the
summary of those models?}
}
\value{
A dataframe indicating the results of the test at each time-bin.
}
\description{
Runs a test on each time-bin of a time-analysis. Supports \code{t.test}, \code{wilcox.test}, \code{lm}, and
\code{lmer}
}

