% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_cluster_data.R
\name{analyze_time_clusters}
\alias{analyze_time_clusters}
\alias{analyze_time_clusters.time_cluster_data}
\title{Bootstrap analysis of time-clusters.}
\usage{
analyze_time_clusters(data, data_options, ...)

\method{analyze_time_clusters}{time_cluster_data}(data, within_subj = FALSE,
  samples = 1000, formula = NULL, shuffle_by = NULL, ...)
}
\arguments{
\item{data}{The output of the \code{make_time_cluster_data} function}

\item{...}{Other args for to selected 'test' function; should be identical to those passed to
\code{make_time_cluster_data} function}

\item{within_subj}{Logical indicating whether to perform within-subjects bootstrap resampling. (Defaults
to FALSE: between-subjects resampling)}

\item{samples}{How many iterations should be performed in the bootstrap resampling procedure?}

\item{formula}{Formula for test. Should be identical to that passed to make_time_cluster_data fxn (if
arg ignored there, can be ignored here)}

\item{shuffle_by}{If the predictor_column is numeric *and* within-subjects, then observations with the
same predictor value could nevertheless correspond to distinct conditions/categories that should be
shuffled separately. For example, when using vocabulary scores to predict looking behavior, a participant
might get identical vocab scores for verbs and nouns; these are nevertheless distinct categories that
should be re-assigned separately when bootstrap-resampling data. The 'shuffle_by' argument allows you to
specify a column which indicates these kinds of distinct categories that should be resampled separately--
but it's only needed if you've specified a numeric *and* within-subjects predictor column.}
}
\value{
A cluster-analysis object, which can be plotted and summarized to examine which temporal periods
  show a significant effect of the predictor variable
}
\description{
Takes data whose time bins have been clustered by significance (using the \code{make_time_cluster_data}
function) and performs a bootstrapping analyses (Maris & Oostenveld, 2007). This analysis takes a summed
statistic for each cluster, and compares it to the "null" distribution of sum statistics obtained by
resampling data within the largest of the clusters.
}
\section{Methods (by class)}{
\itemize{
\item \code{time_cluster_data}: 
}}

