% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{convert_non_aoi_to_trackloss}
\alias{convert_non_aoi_to_trackloss}
\title{Treat gaze data outside of all AOIs as trackloss.}
\usage{
convert_non_aoi_to_trackloss(data, data_options)
}
\arguments{
\item{data}{}

\item{data_options}{}
}
\value{
Data where non-AOI looks are now considered trackloss.
}
\description{
Should gaze outside of any AOI be considered trackloss? This function sets trackloss to TRUE for any
samples that were not inside an AOI
}

