% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{describe_data}
\alias{describe_data}
\title{Describe dataset}
\usage{
describe_data(data, data_options, describe_column, group_columns)
}
\arguments{
\item{data}{}

\item{data_options}{}

\item{describe_column}{The column to return descriptive statistics about.}

\item{group_columns}{Any columns to group by when calculating descriptive statistics (e.g., participants,
conditions, etc.)}
}
\value{
A dataframe giving descriptive statistics for the \code{describe_column}, including mean, SD, var,
min, max, and number of trials
}
\description{
Returns descriptive statistics about a column of choice. A simple wrapper around dplyr:group_by/
dplyr::summarise that allows a quick glance at the data.
}

