% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_threshold}
\alias{.get_threshold}
\title{.get_crit_val()}
\usage{
.get_threshold(alpha, test, dfs, quiet)
}
\arguments{
\item{vec}{Logical}
}
\value{
A numeric vector
}
\description{
Get the threshold, given critical value and test. For `analyze_time_bins`
}
