% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.label_consecutive}
\alias{.label_consecutive}
\title{.label_consecutive()}
\usage{
.label_consecutive(vec)
}
\arguments{
\item{vec}{Logical}
}
\value{
A numeric vector
}
\description{
A helper function to label/enumerate runs of TRUEs in a logical vector, with NA for FALSEs
}
