% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.make_proportion_looking_summary}
\alias{.make_proportion_looking_summary}
\title{.make_proportion_looking_summary()}
\usage{
.make_proportion_looking_summary(data, groups, aoi_col, other_dv_columns)
}
\arguments{
\item{data}{The data}

\item{groups}{The groups}

\item{aoi_col}{Name of AOI column}

\item{other_dv_columns}{Other columns to summarize, aside from AOI column.}
}
\value{
dataframe Transformed dataframe
}
\description{
A helper function for make_time_window_data and make_time_sequence_data. Takes a dataframe, groups it, and
returns proportion looking and relevant transformations
}
