% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_.time_sequence_data}
\alias{filter_.bin_analysis}
\alias{filter_.boot_splines_analysis}
\alias{filter_.boot_splines_data}
\alias{filter_.onset_data}
\alias{filter_.time_cluster_data}
\alias{filter_.time_sequence_data}
\alias{filter_.time_window_data}
\title{EyetrackingR friendly filter}
\usage{
\method{filter_}{time_sequence_data}(data, ...)

\method{filter_}{time_window_data}(data, ...)

\method{filter_}{bin_analysis}(data, ...)

\method{filter_}{boot_splines_data}(data, ...)

\method{filter_}{time_cluster_data}(data, ...)

\method{filter_}{boot_splines_analysis}(data, ...)

\method{filter_}{onset_data}(data, ...)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
EyetrackingR friendly filter
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 

\item \code{time_window_data}: 

\item \code{bin_analysis}: 

\item \code{boot_splines_data}: 

\item \code{time_cluster_data}: 

\item \code{boot_splines_analysis}: 

\item \code{onset_data}: 
}}

