% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_cluster_data.R
\name{get_time_clusters}
\alias{get_time_clusters}
\title{Get information about the clusters in a cluster-analysis}
\usage{
get_time_clusters(object)
}
\arguments{
\item{cluster_analysis}{The output of the \code{analyze_time_clusters} function}
}
\value{
A dataframe with information about the clusters
}
\description{
Get information about the clusters in a cluster-analysis
}

