% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{group_by_.time_sequence_data}
\alias{group_by_.bin_analysis}
\alias{group_by_.boot_splines_analysis}
\alias{group_by_.boot_splines_data}
\alias{group_by_.onset_data}
\alias{group_by_.time_cluster_data}
\alias{group_by_.time_sequence_data}
\alias{group_by_.time_window_data}
\title{EyetrackingR friendly group_by}
\usage{
\method{group_by_}{time_sequence_data}(data, ...)

\method{group_by_}{time_window_data}(data, ...)

\method{group_by_}{bin_analysis}(data, ...)

\method{group_by_}{boot_splines_data}(data, ...)

\method{group_by_}{time_cluster_data}(data, ...)

\method{group_by_}{boot_splines_analysis}(data, ...)

\method{group_by_}{onset_data}(data, ...)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
EyetrackingR friendly group_by
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 

\item \code{time_window_data}: 

\item \code{bin_analysis}: 

\item \code{boot_splines_data}: 

\item \code{time_cluster_data}: 

\item \code{boot_splines_analysis}: 

\item \code{onset_data}: 
}}

