% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{keep_trackloss}
\alias{keep_trackloss}
\title{Keep data with trackloss in dataset.}
\usage{
keep_trackloss(data, data_options)
}
\arguments{
\item{data}{}

\item{data_options}{}
}
\value{
Data with trackloss preserved
}
\description{
Converts data so that, in all subsequent proportion-looking calculations, proportion_looking_at_aoi =
time_looking_at_aoi / time_possible. Trackloss is *not* considered missing data, but instead is counted as
time *not* spent looking at each AOI.
}

