% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_spline_data.R
\name{make_boot_splines_data}
\alias{make_boot_splines_data}
\title{Bootstrap resample splines for time-series data}
\usage{
make_boot_splines_data(data, ...)
}
\arguments{
\item{data}{}

\item{predictor_column}{What predictor var to split by? Maximum two conditions}

\item{within_subj}{Are the two conditions within or between subjects?}

\item{samples}{How many (re)samples to take?}

\item{resolution}{What resolution should we return predicted splines at, in ms? e.g., 10ms = 100 intervals
per second, or hundredths of a second}

\item{alpha}{p-value when the groups are sufficiently "diverged"}

\item{smoother}{Smooth data using "smooth.spline," "loess," or leave NULL for no smoothing}
}
\value{
A bootstrapped distribution of samples for each time-bin
}
\description{
Bootstrap splines from \code{time_sequence_data()}. This creates a distribution from which a non-parametric
analysis can be performed.
}

