% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{make_eyetrackingr_data}
\alias{make_eyetrackingr_data}
\title{Convert raw data for use in eyetrackingR}
\usage{
make_eyetrackingr_data(data, participant_column, trackloss_column, time_column,
  trial_column, aoi_columns, treat_non_aoi_looks_as_missing,
  item_columns = NULL)
}
\arguments{
\item{data}{}

\item{participant_column}{Column name for participant identifier}

\item{trackloss_column}{Column name indicating trackloss}

\item{time_column}{Column name indicating time}

\item{trial_column}{Column name indicating trial identifier}

\item{aoi_columns}{Names of AOIs}

\item{treat_non_aoi_looks_as_missing}{This is a logical indicating how you would like to
perform "proportion-looking" calculations, which are critical to any eyetracking analysis. If
set to TRUE, any samples that are not in any of the AOIs (defined with the \code{aoi_columns}
argument) are treated as missing data; when it comes time for eyetrackingR to calculate
proportion looking to an AOI, this will be calculated as "time looking to that AOI divided by
time looking to all other AOIs." In contrast, if this parameter is set to FALSE, proportion
looking to an AOI will be calculated as "time looking to that AOI divided by total time
looking."}

\item{item_columns}{Column names indicating items (optional)}
}
\value{
Dataframe ready for use in eyetrackingR.
}
\description{
This should be the first function you use when using eyetrackingR for a project.
}
\details{
This takes your raw dataframe, as well as a list of information about your dataframe. It confirms
that all the columns are the right format, and outputs a dataframe that is ready to used with
eyetrackingR's functions.
}

