% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onset_contingent_data.R
\name{make_onset_data}
\alias{make_onset_data}
\title{Make onset-contingent data.}
\usage{
make_onset_data(data, data_options, onset_time, fixation_window_length,
  target_aoi, distractor_aoi = NULL)
}
\arguments{
\item{data}{The original (verified) data}

\item{data_options}{}

\item{onset_time}{When to check for participants' "starting" AOI?}

\item{fixation_window_length}{Which AOI is currently being fixated is determined by taking a rolling
average. This is the width of window for rolling average.}

\item{target_aoi}{Which AOI is the target that should be switched *to*}

\item{distractor_aoi}{Which AOI is the distractor that should be switched *from* (default = !target_aoi)}
}
\value{
Original dataframe augmented with column indicating switch away from target AOI
}
\description{
Divide trials into which AOI participants started on. Calculate switches away from this AOI, using
a rolling window to determine what length consitutes a switch. Augment original data with a column
indicating whether each row is a switch-away sample.
}

