% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/onset_contingent_data.R
\name{make_switch_data}
\alias{make_switch_data}
\alias{make_switch_data.onset_data}
\title{Summarize data into time-to-switch from initial AOI.}
\usage{
make_switch_data(data, ...)

\method{make_switch_data}{onset_data}(data, predictor_columns = NULL,
  summarize_by = NULL)
}
\arguments{
\item{data}{The output of \code{make_onset_data}}

\item{predictor_columns}{Variables/covariates of interest when analyzing time-to-switch}

\item{summarize_by}{Should the data be summarized along, e.g., participants, items, etc.? If so, give
column name(s) here. If left blank, will leave trials distinct. The former is needed for more traditional
analyses (t.tests, ANOVAs), while the latter is preferable for mixed-effects models (lmer)}
}
\value{
A dataframe indicating initial AOI and time-to-switch from that AOI for each trial/subject/item/etc.
}
\description{
Take trials split by initial-AOI, and determine how quickly participants switch away from that AOI
}
\section{Methods (by class)}{
\itemize{
\item \code{onset_data}: 
}}

