% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_sequence_data.R
\name{make_time_sequence_data}
\alias{make_time_sequence_data}
\title{make_time_sequence_data()}
\usage{
make_time_sequence_data(data, data_options, time_bin_size, aois = NULL,
  predictor_columns = NULL, summarize_by = NULL)
}
\arguments{
\item{data}{}

\item{data_options}{}

\item{time_bin_size}{}

\item{aois}{Which AOIs are of interest? Defaults to all in 'data_options'}

\item{predictor_columns}{Which columns indicate predictor vars, and therefore should be preserved in
grouping operations?}

\item{summarize_by}{Should the data be summarized along, e.g., participants, items, etc.? If so, give
  column name(s) here. If left blank, will leave trials distinct. The former is needed for more traditional
  analyses (t.tests, ANOVAs), while the latter is preferable for mixed-effects models (lmer)}
}
\value{
Data binned into time-bins, with proportion-looking and transformations as well as orthogonal
  time-polynomials for growth curve analysis
}
\description{
Creates time-bins and summarizes proportion-looking within each time-bin
}

