% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_window_data.R
\name{make_time_window_data}
\alias{make_time_window_data}
\title{Make a dataset collapsing over a time-window}
\usage{
make_time_window_data(data, aois = NULL, predictor_columns = NULL,
  summarize_by = NULL)
}
\arguments{
\item{data}{}

\item{aois}{Which AOIs are of interest? Defaults to all in 'data_options'}

\item{predictor_columns}{Which columns indicate predictor vars, and therefore should be preserved in
grouping operations?}

\item{summarize_by}{Should the data be summarized along, e.g., participants, items, etc.? If so, give
column names here. If left blank, will leave trials distinct. The former is needed for more traditional
analyses (t.tests, ANOVAs), while the latter is preferable for mixed-effects models (lmer)}
}
\value{
Data with proportion-looking and transformations (empirical logit, arc-sin)
}
\description{
Collapse time across our entire window and return a dataframe ready for analyses
}

