% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{mutate_.time_sequence_data}
\alias{mutate_.bin_analysis}
\alias{mutate_.boot_splines_analysis}
\alias{mutate_.boot_splines_data}
\alias{mutate_.onset_data}
\alias{mutate_.time_cluster_data}
\alias{mutate_.time_sequence_data}
\alias{mutate_.time_window_data}
\title{EyetrackingR friendly mutate}
\usage{
\method{mutate_}{time_sequence_data}(data, ...)

\method{mutate_}{time_window_data}(data, ...)

\method{mutate_}{bin_analysis}(data, ...)

\method{mutate_}{boot_splines_data}(data, ...)

\method{mutate_}{time_cluster_data}(data, ...)

\method{mutate_}{boot_splines_analysis}(data, ...)

\method{mutate_}{onset_data}(data, ...)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
EyetrackingR friendly mutate
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 

\item \code{time_window_data}: 

\item \code{bin_analysis}: 

\item \code{boot_splines_data}: 

\item \code{time_cluster_data}: 

\item \code{boot_splines_analysis}: 

\item \code{onset_data}: 
}}

