% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_spline_data.R
\name{plot.boot_splines_analysis}
\alias{plot.boot_splines_analysis}
\title{Plot differences in bootstrapped-splines data}
\usage{
\method{plot}{boot_splines_analysis}(x, ...)
}
\arguments{
\item{x}{The output of the \code{analyze_boot_splines} function}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Plot the means and CIs of bootstrapped spline difference estimates and intervals
(either within-subjects or between-subjects)
}

