% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spline_data.R
\name{plot.boot_splines_data}
\alias{plot.boot_splines_data}
\title{Plot bootstrapped-splines data}
\usage{
\method{plot}{boot_splines_data}(x, ...)
}
\arguments{
\item{x}{The output of the \code{make_boot_splines_data} function}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Plot the means and CIs of bootstrapped splines (either within-subjects or between-subjects)
}
