% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_cluster_data.R
\name{plot.cluster_analysis}
\alias{plot.cluster_analysis}
\title{Visualize the results of a cluster analysis.}
\usage{
\method{plot}{cluster_analysis}(cl_analysis)
}
\arguments{
\item{cluster_analysis}{object returned by cluster_analysis()}
}
\value{
A ggplot object
}
\description{
Plots the result of the bootstrapping cluster analysis. A histogram of the sum statistics for the
shuffled (null) distribution, with the sum statisics for each of the clusters indicated by dashed lines.
}

