% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot.data.frame}
\alias{plot.data.frame}
\title{Prevent accidental plotting of data.frames whose class has been removed.}
\usage{
\method{plot}{data.frame}(data, ...)
}
\arguments{
\item{data}{}
}
\description{
EyetrackingR adds a plot method to most dataframes you can create with it.
Sometimes, the class of the dataframe can be removed, so that plot will
not have the expected result. This ensures a warning is issued to the user.
}

