% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{plot.eyetrackingR_data_summary}
\alias{plot.eyetrackingR_data_summary}
\title{Plot some summarized data from eyetrackingR}
\usage{
\method{plot}{eyetrackingR_data_summary}(data_summary)
}
\arguments{
\item{data}{The data returned by make_time_window_data()}
}
\value{
A ggplot object
}
\description{
Plots the data returned from \code{describe_data}. Like that function, this is a convenient
wrapper good for sanity checks.
#'
}

