% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset_contingent_data.R
\name{plot.onset_data}
\alias{plot.onset_data}
\title{Plot onset-contingent data}
\usage{
\method{plot}{onset_data}(x, predictor_columns = NULL, ...)
}
\arguments{
\item{x}{The output of the \code{make_onset_data} function}

\item{predictor_columns}{Column(s) by which to facet the data. Maximum two columns. Will perform
median split if numeric.}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Divide trials into which AOI participants started on; plot proportion looking away from that AOI.
}

