% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onset_contingent_data.R
\name{plot.switch_data}
\alias{plot.switch_data}
\title{Plot mean switch-from-initial-AOI times.}
\usage{
\method{plot}{switch_data}(x, predictor_columns = NULL, ...)
}
\arguments{
\item{x}{The output of the \code{make_switch_data} function}

\item{predictor_columns}{Column(s) by which to facet the data. Maximum two columns. Will perform
median split if numeric.}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Boxplot of mean switch time aggregated by subjects within each FirstAOI, potentially faceted by
predictor_columns.
}
