% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_cluster_data.R
\name{print.cluster_analysis}
\alias{print.cluster_analysis}
\title{Print Method for Cluster Analysis}
\usage{
\method{print}{cluster_analysis}(x, ...)
}
\arguments{
\item{x}{The output of the \code{analyze_clusters} function}

\item{...}{Ignored}
}
\value{
Prints information about the bootstrapped null distribution, as well as information about each cluster.
}
\description{
Print Method for Cluster Analysis
}

