% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{remove_trackloss}
\alias{remove_trackloss}
\title{Remove data with trackloss from dataset.}
\usage{
remove_trackloss(data, data_options, delete_rows = FALSE)
}
\arguments{
\item{data}{}

\item{data_options}{}

\item{delete_rows}{(default: FALSE) Should rows with trackloss be deleted, or simply set to NA?}
}
\value{
Data with trackloss removed
}
\description{
Converts data so that, in all subsequent proportion-looking calculations, proportion_looking_at_aoi =
time_looking_at_aoi / time_looking. Trackloss *is* considered missing data, and time spent not looking at a
given AOI must have been spent looking at another AOI
}

