% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{set_data_options}
\alias{set_data_options}
\title{Set data options.}
\usage{
set_data_options(participant_column, trackloss_column, time_column,
  trial_column, item_columns = NULL, aoi_columns)
}
\arguments{
\item{participant_column}{Column name for participant identifier}

\item{trackloss_column}{Column name indicating trackloss}

\item{time_column}{Column name indicating time}

\item{trial_column}{Column name indicating trial identifier}

\item{item_columns}{Column names indicating items (can be same as trial_column)}

\item{aoi_columns}{Names of AOIs}
}
\value{
list of configuration options
}
\description{
Create a list which describes the important aspects of your data, to be used by most other functions in
this package.
}

