% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simulate_eyetrackingr_data}
\alias{simulate_eyetrackingr_data}
\title{Simulate an eyetrackingR dataset}
\usage{
simulate_eyetrackingr_data(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Prints a list of divergence-times.
}
\description{
This function creates an eyetrackingR dataset (i.e., already run through make_eyetrackingr_data).
This can be helpful for examining the false-alarm and sensitivity of analysis-techniques via
simulations.
}
\details{
The data is an experiment with 6 items and two (within-subjects, across-items) conditions.

There are two AOIs, and \code{treat_non_aoi_looks_as_missing = TRUE}.
}

