% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{subset.time_sequence_data}
\alias{subset.bin_analysis}
\alias{subset.boot_splines_analysis}
\alias{subset.boot_splines_data}
\alias{subset.onset_data}
\alias{subset.time_cluster_data}
\alias{subset.time_sequence_data}
\alias{subset.time_window_data}
\title{EyetrackingR friendly subset}
\usage{
\method{subset}{time_sequence_data}(data, ...)

\method{subset}{time_window_data}(data, ...)

\method{subset}{bin_analysis}(data, ...)

\method{subset}{boot_splines_data}(data, ...)

\method{subset}{time_cluster_data}(data, ...)

\method{subset}{boot_splines_analysis}(data, ...)

\method{subset}{onset_data}(data, ...)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe
}
\description{
EyetrackingR friendly subset
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 

\item \code{time_window_data}: 

\item \code{bin_analysis}: 

\item \code{boot_splines_data}: 

\item \code{time_cluster_data}: 

\item \code{boot_splines_analysis}: 

\item \code{onset_data}: 
}}

