% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{subset_by_window}
\alias{subset_by_window}
\title{Extract a subset of the dataset within a time-window in each trial.}
\usage{
subset_by_window(data, data_options, window_start = -Inf, window_end = Inf,
  rezero = NULL)
}
\arguments{
\item{data}{}

\item{data_options}{}

\item{window_start}{Number (for timestamp) or character (for column that specifies timestamp)}

\item{window_end}{Number (for timestamp) or character (for column that specifies timestamp)}

\item{rezero}{Should the beginning of the window be considered the zero point of the timestamp?
Default TRUE when window_start is column, FALSE when window_start is number}
}
\value{
Subsetted data
}
\description{
Extract a subset of the dataset, where each trial falls inside a time-window.
Time-window can either be specifed by a number (for a timestamp across all trials) or
by a column which picks out a timestamp for each participant/trial
}

