% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_sequence_data.R
\name{summary.bin_analysis}
\alias{summary.bin_analysis}
\title{Summary Method for Time-bin Analysis}
\usage{
\method{summary}{bin_analysis}(object, ...)
}
\arguments{
\item{object}{The output of the \code{analyze_time_bins} function}

\item{...}{Ignored}
}
\value{
Prints information about each run of statistically significant time-bins, separately for
  positive and negative
}
\description{
Summary Method for Time-bin Analysis
}

