% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spline_data.R
\name{summary.boot_splines_analysis}
\alias{summary.boot_splines_analysis}
\title{Summary Method for Bootstrapped Splines Analysis}
\usage{
\method{summary}{boot_splines_analysis}(object, ...)
}
\arguments{
\item{object}{The output of the \code{boot_splines_data} function}

\item{...}{Ignored}
}
\value{
Prints a list of divergence-times.
}
\description{
Summary Method for Bootstrapped Splines Analysis
}

