% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_cluster_data.R
\name{summary.time_cluster_data}
\alias{summary.time_cluster_data}
\title{Summary Method for Cluster Analysis}
\usage{
\method{summary}{time_cluster_data}(object, ...)
}
\arguments{
\item{object}{The output of the \code{analyze_clusters} function}

\item{...}{Ignored}
}
\value{
Prints information about the bootstrapped null distribution, as well as information about each cluster.
}
\description{
Summary Method for Cluster Analysis
}

