% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{trackloss_analysis}
\alias{trackloss_analysis}
\title{Analyze trackloss.}
\usage{
trackloss_analysis(data)
}
\arguments{
\item{data}{}
}
\value{
A dataframe describing trackloss by-trial and by-participant
}
\description{
Get information on trackloss in your data.
}

