% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{ungroup.time_sequence_data}
\alias{ungroup.bin_analysis}
\alias{ungroup.boot_splines_analysis}
\alias{ungroup.boot_splines_data}
\alias{ungroup.eyetrackingR}
\alias{ungroup.onset_data}
\alias{ungroup.time_cluster_data}
\alias{ungroup.time_sequence_data}
\alias{ungroup.time_window_data}
\title{EyetrackingR friendly ungroup}
\usage{
\method{ungroup}{time_sequence_data}(x)

\method{ungroup}{time_window_data}(x)

\method{ungroup}{bin_analysis}(x)

\method{ungroup}{boot_splines_data}(x)

\method{ungroup}{time_cluster_data}(x)

\method{ungroup}{boot_splines_analysis}(x)

\method{ungroup}{onset_data}(x)

\method{ungroup}{eyetrackingR}(x)
}
\arguments{
\item{x}{An eyetrackingR dataframe}
}
\value{
An eyetrackingR dataframe, without dplyr groups
}
\description{
EyetrackingR friendly ungroup
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 

\item \code{time_window_data}: 

\item \code{bin_analysis}: 

\item \code{boot_splines_data}: 

\item \code{time_cluster_data}: 

\item \code{boot_splines_analysis}: 

\item \code{onset_data}: 

\item \code{eyetrackingR}: 
}}

