% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_prep.R
\name{verify_dataset}
\alias{verify_dataset}
\title{Verify the dataset}
\usage{
verify_dataset(data, data_options)
}
\arguments{
\item{data}{Your data}

\item{data_options}{Created by \code{set_data_options}}
}
\value{
Dataset with verified column types.
}
\description{
Use data_options to verify that the columns in your dataset are of the correct type
}

