% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eyetrackingR.R
\docType{package}
\name{eyetrackingR}
\alias{eyetrackingR}
\alias{eyetrackingR-package}
\title{eyetrackingR: A package for cleaning, analyzing, and visualizing eye-tracking datasets}
\description{
This package addresses tasks along the pipeline from raw eye-tracking data to analysis and
visualization. It offers several popular types of analyses, including linear and
growth curve time analyses, onset-contingent reaction time analyses, and cluster mass
analyses, as well as novel non-parametric approaches to time-series data.
}
\details{
For more information and tutorials, visit \href{http://www.eyetrackingr.com}{eyetrackingR.com}.
}

