% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_.time_sequence_data}
\alias{filter_.bin_analysis}
\alias{filter_.boot_splines_analysis}
\alias{filter_.boot_splines_data}
\alias{filter_.eyetrackingR}
\alias{filter_.onset_data}
\alias{filter_.time_cluster_data}
\alias{filter_.time_sequence_data}
\alias{filter_.time_window_data}
\title{EyetrackingR friendly filter}
\usage{
\method{filter_}{time_sequence_data}(.data, ..., .dots)

\method{filter_}{time_window_data}(.data, ..., .dots)

\method{filter_}{bin_analysis}(.data, ..., .dots)

\method{filter_}{boot_splines_data}(.data, ..., .dots)

\method{filter_}{time_cluster_data}(.data, ..., .dots)

\method{filter_}{boot_splines_analysis}(.data, ..., .dots)

\method{filter_}{onset_data}(.data, ..., .dots)

\method{filter_}{eyetrackingR}(.data, ..., .dots)
}
\arguments{
\item{.data}{An eyetrackingR dataframe}

\item{...}{Logical predicates. Multiple conditions are combined with &.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse") for details.}
}
\value{
A dataframe
}
\description{
EyetrackingR friendly filter
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: eyetrackingR friendly filter

\item \code{time_window_data}: eyetrackingR friendly filter

\item \code{bin_analysis}: eyetrackingR friendly filter

\item \code{boot_splines_data}: eyetrackingR friendly filter

\item \code{time_cluster_data}: eyetrackingR friendly filter

\item \code{boot_splines_analysis}: eyetrackingR friendly filter

\item \code{onset_data}: eyetrackingR friendly filter

\item \code{eyetrackingR}: eyetrackingR friendly filter
}}

