% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{group_by_.time_sequence_data}
\alias{group_by_.bin_analysis}
\alias{group_by_.boot_splines_analysis}
\alias{group_by_.boot_splines_data}
\alias{group_by_.eyetrackingR}
\alias{group_by_.onset_data}
\alias{group_by_.time_cluster_data}
\alias{group_by_.time_sequence_data}
\alias{group_by_.time_window_data}
\title{EyetrackingR friendly group_by}
\usage{
\method{group_by_}{time_sequence_data}(.data, ..., .dots, add = FALSE)

\method{group_by_}{time_window_data}(.data, ..., .dots, add = FALSE)

\method{group_by_}{bin_analysis}(.data, ..., .dots, add = FALSE)

\method{group_by_}{boot_splines_data}(.data, ..., .dots, add = FALSE)

\method{group_by_}{time_cluster_data}(.data, ..., .dots, add = FALSE)

\method{group_by_}{boot_splines_analysis}(.data, ..., .dots, add = FALSE)

\method{group_by_}{onset_data}(.data, ..., .dots, add = FALSE)

\method{group_by_}{eyetrackingR}(.data, ..., .dots, add = FALSE)
}
\arguments{
\item{.data}{An eyetrackingR dataframe}

\item{...}{variables to group by. Duplicated groups will be silently dropped.}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse") for details.}

\item{add}{By default, when add = FALSE, group_by will override existing groups. To instead add
to the existing groups, use add = TRUE}
}
\value{
A dataframe
}
\description{
EyetrackingR friendly group_by
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 

\item \code{time_window_data}: 

\item \code{bin_analysis}: 

\item \code{boot_splines_data}: 

\item \code{time_cluster_data}: 

\item \code{boot_splines_analysis}: 

\item \code{onset_data}: 

\item \code{eyetrackingR}: 
}}

