% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_spline_data.R
\name{make_boot_splines_data}
\alias{make_boot_splines_data}
\alias{make_boot_splines_data.time_sequence_data}
\title{Bootstrap resample splines for time-series data.}
\usage{
make_boot_splines_data(data, predictor_column, aoi, within_subj, smoother,
  samples, resolution, alpha)

\method{make_boot_splines_data}{time_sequence_data}(data, predictor_column,
  aoi = NULL, within_subj = FALSE, smoother = "smooth.spline",
  samples = 1000, resolution = NULL, alpha = 0.05)
}
\arguments{
\item{data}{The output of \code{time_sequence_data()}}

\item{predictor_column}{What predictor var to split by? Maximum two conditions}

\item{aoi}{Which AOI do you wish to perform the analysis on?}

\item{within_subj}{Are the two conditions within or between subjects?}

\item{smoother}{Smooth data using "smooth.spline," "loess," or leave NULL for no smoothing}

\item{samples}{How many iterations to run bootstrap resampling? Default 1000}

\item{resolution}{What resolution should we return predicted splines at, in ms? e.g., 10ms = 100
intervals per second, or hundredths of a second. Default is the same size as time-bins.}

\item{alpha}{p-value when the groups are sufficiently "diverged"}
}
\value{
A bootstrapped distribution of samples for each time-bin
}
\description{
Bootstrap splines from \code{time_sequence_data()}. This creates a distribution from which a non-parametric
analysis can be performed.
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 
}}

