% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_cluster_data.R
\name{make_time_cluster_data}
\alias{make_time_cluster_data}
\alias{make_time_cluster_data.time_sequence_data}
\title{Make data for cluster analysis.}
\usage{
make_time_cluster_data(data, ...)

\method{make_time_cluster_data}{time_sequence_data}(data, predictor_column,
  aoi = NULL, test, threshold = NULL, formula = NULL, ...)
}
\arguments{
\item{data}{The output of the \code{make_time_sequence_data} function}

\item{...}{Any other arguments to be passed to the selected 'test' function (e.g., paired,
var.equal, etc.)}

\item{predictor_column}{The variable whose test statistic you are interested in. Testing the
intercept and interaction terms not currently supported}

\item{aoi}{If this dataframe has multiple AOIs, you must specify which to analyze}

\item{test}{What type of test should be performed in each time bin? Supports
\code{t.test}, \code{wilcox}, \code{lm}, or \code{lmer}.}

\item{threshold}{Value of statistic used in determining significance. Note the sign! This
test is directional.}

\item{formula}{What formula should be used for test? Optional (for all but \code{lmer}), if unset
uses \code{Prop ~ [predictor_column]}}
}
\value{
The original data, augmented with information about clusters. Calling summary on this data will
  describe these clusters. The dataset is ready for the \code{\link{analyze_time_clusters}} method.
}
\description{
Takes data that has been summarized into time-bins by \code{make_time_sequence_data()}, finds adjacent time
bins that pass some threshold of significance, and assigns these adjacent bins into groups (clusters).
Output is ready for a bootstrapping analyses (Maris & Oostenveld, 2007)
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 
}}

