% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mutate_.time_sequence_data}
\alias{mutate_.bin_analysis}
\alias{mutate_.boot_splines_analysis}
\alias{mutate_.boot_splines_data}
\alias{mutate_.eyetrackingR}
\alias{mutate_.onset_data}
\alias{mutate_.time_cluster_data}
\alias{mutate_.time_sequence_data}
\alias{mutate_.time_window_data}
\title{EyetrackingR friendly mutate}
\usage{
\method{mutate_}{time_sequence_data}(.data, ..., .dots)

\method{mutate_}{time_window_data}(.data, ..., .dots)

\method{mutate_}{bin_analysis}(.data, ..., .dots)

\method{mutate_}{boot_splines_data}(.data, ..., .dots)

\method{mutate_}{time_cluster_data}(.data, ..., .dots)

\method{mutate_}{boot_splines_analysis}(.data, ..., .dots)

\method{mutate_}{onset_data}(.data, ..., .dots)

\method{mutate_}{eyetrackingR}(.data, ..., .dots)
}
\arguments{
\item{.data}{An eyetrackingR dataframe}

\item{...}{name value pairs of expressions}

\item{.dots}{Used to work around non-standard evaluation. See vignette("nse") for details.}
}
\value{
A dataframe
}
\description{
EyetrackingR friendly mutate
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: eyetrackingR friendly mutate

\item \code{time_window_data}: eyetrackingR friendly mutate

\item \code{bin_analysis}: eyetrackingR friendly mutate

\item \code{boot_splines_data}: eyetrackingR friendly mutate

\item \code{time_cluster_data}: eyetrackingR friendly mutate

\item \code{boot_splines_analysis}: eyetrackingR friendly mutate

\item \code{onset_data}: eyetrackingR friendly mutate

\item \code{eyetrackingR}: eyetrackingR friendly mutate
}}

