% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_sequence_data.R
\name{plot.bin_analysis}
\alias{plot.bin_analysis}
\title{Plot test-statistic for each time-bin in a time-series}
\usage{
\method{plot}{bin_analysis}(x, type = NULL, ...)
}
\arguments{
\item{x}{The output of \code{analyze_time_bins}}

\item{type}{This function can plot the test-statistic ("statistic"), the parameter estimate +/-
std. error ("estimate"), the p-value ("pvalue") or the negative-log-pvalue ("neg_log_pvalue").
When test gives critical-statistic, default is to plot the test-statistic. Otherwise, default
is to plot the estimate. For wilcox, only p-values can be plotted.}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Plot the result from the \code{analyze_time_bins} function, with the statistic and threshold for each bin
}
