% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_sequence_data.R
\name{plot.bin_analysis}
\alias{plot.bin_analysis}
\title{Plot test-statistic for each time-bin in a time-series}
\usage{
\method{plot}{bin_analysis}(x, type = "statistic", ...)
}
\arguments{
\item{x}{The output of \code{analyze_time_bins}}

\item{type}{Plot the test-statistic at each time bin ("statistic"), or the parameter estimate at each time
bin ("estimate")? Note that estimate plots standard error, regardless of the alpha/threshold used originally}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Plot the result from the \code{analyze_time_bins} function, with the statistic and threshold for each bin
}

