% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_cluster_data.R
\name{plot.time_cluster_data}
\alias{plot.time_cluster_data}
\title{Plot test-statistic for each time-bin in a time-series, highlight clusters.
Plot time_cluster_data, highlights clusters of above-threshold time-bins.}
\usage{
\method{plot}{time_cluster_data}(x, clusters = NULL, ...)
}
\arguments{
\item{x}{The output of \code{make_time_cluster_data}}

\item{clusters}{A vector of the clusters you'd like highlighted. If left blank, all are highlighted}

\item{...}{Ignored}
}
\value{
A ggplot object
}
\description{
Plot test-statistic for each time-bin in a time-series, highlight clusters.
Plot time_cluster_data, highlights clusters of above-threshold time-bins.
}

