% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time_sequence_data.R
\name{plot.time_sequence_data}
\alias{plot.time_sequence_data}
\title{Plot time-sequence data}
\usage{
\method{plot}{time_sequence_data}(data, predictor_column = NULL,
  dv = "Prop")
}
\arguments{
\item{data}{}

\item{predictor_column}{}
}
\value{
A ggplot object
}
\description{
Plot the timecourse of looking. Each AOI will be plotted in a separate pane, and data can be split into
groups by a predictor column (median split if numeric).
}

