% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reclass}
\alias{reclass}
\alias{reclass.eyetrackingR_df}
\title{Add the original class/attributes back onto result (usually of dplyr operation)}
\usage{
reclass(x, result, ...)

\method{reclass}{eyetrackingR_df}(x, result, ...)
}
\arguments{
\item{x}{The original object, class information you want to restore.}

\item{result}{Some transformation of \code{x}, which may have removed its class/attributes.}

\item{...}{Ignored}
}
\value{
The \code{result}, now with class/attribute information restored.

The \code{result}, now with class/attribute information restored.
}
\description{
Add the original class/attributes back onto result (usually of dplyr operation)
}
\section{Methods (by class)}{
\itemize{
\item \code{reclass(eyetrackingR_df)}: Add the original class/attributes back onto result (usually of dplyr operation)

}}
