% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simulate_eyetrackingr_data}
\alias{simulate_eyetrackingr_data}
\title{Simulate an eyetrackingR dataset}
\usage{
simulate_eyetrackingr_data(num_trials_per_condition = 1, ...)
}
\arguments{
\item{num_trials_per_condition}{Number of trials per-subject per-condition.}

\item{...}{Ignored}
}
\value{
Prints a list of divergence-times.
}
\description{
This function creates an eyetrackingR dataset (i.e., already run through make_eyetrackingr_data).
This can be helpful for examining the false-alarm and sensitivity of analysis-techniques via
simulations.
}

