% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{subset.time_sequence_data}
\alias{subset.bin_analysis}
\alias{subset.boot_splines_analysis}
\alias{subset.boot_splines_data}
\alias{subset.eyetrackingR}
\alias{subset.onset_data}
\alias{subset.time_cluster_data}
\alias{subset.time_sequence_data}
\alias{subset.time_window_data}
\title{EyetrackingR friendly subset}
\usage{
\method{subset}{time_sequence_data}(x, ...)

\method{subset}{time_window_data}(x, ...)

\method{subset}{bin_analysis}(x, ...)

\method{subset}{boot_splines_data}(x, ...)

\method{subset}{time_cluster_data}(x, ...)

\method{subset}{boot_splines_analysis}(x, ...)

\method{subset}{onset_data}(x, ...)

\method{subset}{eyetrackingR}(x, ...)
}
\arguments{
\item{x}{A dataframe to be subsetted}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A dataframe with only selected elements
}
\description{
EyetrackingR friendly subset
}
\section{Methods (by class)}{
\itemize{
\item \code{time_sequence_data}: 

\item \code{time_window_data}: 

\item \code{bin_analysis}: 

\item \code{boot_splines_data}: 

\item \code{time_cluster_data}: 

\item \code{boot_splines_analysis}: 

\item \code{onset_data}: 

\item \code{eyetrackingR}: 
}}

